wip reminder email worker  in tdd
