+++ backend/app/workers/reminder_email_worker.rb
@@ -5,21 +5,9 @@ require "sidekiq-scheduler"
 class ReminderEmailWorker
   include Sidekiq::Worker
 
-  attr_reader :enrollment
-
-  def initialize(enrollments)
-    @enrollment = enrollments
-  end
-
   def perform(user_id)
     # @enrollment_list = draft_enrollments
   end
-
-  def draft_enrollments(enrollments)
-    Enrollment.where(status: "draft")
-      .where({updated_at: (Time.now.midnight - 1.months)..Time.now.midnight - 15.days})
-  end
-
   # Only send an email to demandeurs if enrollment are in draft for more than 15 days without changes.
   def send?(enrollment)
   end
